Create dedicated controllers and services for each calculation model (#1033)
---------
